test: Corrected imports in tests after refactoring position
